Mopho 2: Created constants_GUI_PathData.h and added path data for file and folder icons and the pixels that form the LCD characters for the program name display. Created gui_build_LCD_CharacterPath.h and declared and defined struct LCD_CharacterPath and its single static member function buildForChar(). Created 1_comp_KnobAndAttachment_ForVoiceNameChar.h and .cpp and declared and defined class KnobAndAttachment_ForVoiceNameChar. In ExposedParamControl, included an instance of KnobAndAttachment_ForVoiceNameChar and declared and defined buildKnobAndAttachment_ForVoiceNameChar_ControlForExposedParam. Renamed buildKnobAndAttachmentControl_ForOscShape_ForExposedParam() to buildKnobAndAttachment_ForOscShape_ControlForExposedParam(). Renamed buildKnobAndAttachmentControl_ForSeqStep_ForExposedParam() to buildKnobAndAttachment_ForSeqStep_ControlForExposedParam(). Renamed buildToggleButtonAndAttachment_ForExposedParam() to buildToggleButtonAndAttachmentControlForExposedParam(). Renamed buildComboBoxAndAttachment_ForExposedParam() to buildComboBoxAndAttachmentControlForExposedParam(). In DescriptionString, declared and defined buildFor_VoiceNameChar(). In InfoForExposedParameters, added description properties for all the voice name characters. Modified ExposedParamControlsServer::fillControlsVector() to include all of the exposed parameters.

+++ MophoGUI/MophoGUI 2.0/Source/exposedParamControls/sliders/1_comp_KnobAndAttachment_ForSeqStep.cpp
@@ -91,7 +91,7 @@ void KnobAndAttachment_ForSeqStep::setKnobIsNotModifyingPitch() {
 void KnobAndAttachment_ForSeqStep::parameterValueChanged(int changedParamIndex, float newValue) {
 	if (changedParamIndex == trackDestIndex || changedParamIndex == paramIndex) {
 		auto& info{ InfoForExposedParameters::get() };
-		auto paramID{ info.IDfor(changedParamIndex) };
+		auto paramID{ info.IDfor((uint8)changedParamIndex) };
 		auto paramaterPtr{ exposedParams->getParameter(paramID) };
 		auto currentChoice{ roundToInt(paramaterPtr->convertFrom0to1(newValue)) };
 		if (changedParamIndex == trackDestIndex) {
